Public release

+++ RadioDns-PlugIt/radiovis/actions.py
@@ -1,23 +1,27 @@
 # -*- coding: utf-8 -*-
 
 # Utils
-from plugit.utils import action, only_orga_member_user, only_orga_admin_user, PlugItRedirect, json_only, PlugItSendFile, addressInNetwork
+from plugit.utils import action, only_orga_member_user, only_orga_admin_user, PlugItRedirect, json_only, PlugItSendFile, \
+    addressInNetwork, no_template
+from werkzeug.utils import secure_filename
 
-from models import db, Station, Channel, Picture, Ecc, LogEntry
-
+from models import db, Station, Channel, Picture, Ecc, LogEntry, ServiceProvider
+from plugit.api import PlugItAPI, Orga
 import urlparse
+from aws import awsutils
 
 import os
 import sys
 import time
+import uuid
 
-from werkzeug import secure_filename
 from flask import abort
 
 from PIL import Image
 import imghdr
 
 import config
+from utils import send_image_to_mock_api
 
 
 @action(route="/radiovis/gallery/", template="radiovis/gallery/home.html")
@@ -27,7 +31,7 @@ def radiovis_gallery_home(request):
 
     list = []
 
-    for elem in Picture.query.filter_by(orga = int(request.args.get('ebuio_orgapk'))).order_by(Picture.name).all():
+    for elem in Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk'))).order_by(Picture.name).all():
         list.append(elem.json)
 
     saved = request.args.get('saved') == 'yes'
@@ -44,13 +48,24 @@ def radiovis_gallery_edit(request, id):
     object = None
     errors = []
 
+    plugitapi = PlugItAPI(config.API_URL)
+    orga = plugitapi.get_orga(request.args.get('ebuio_orgapk') or request.form.get('ebuio_orgapk'))
+
+    sp = None
+    if orga.codops:
+        sp = ServiceProvider.query.filter_by(codops=orga.codops).order_by(ServiceProvider.codops).first()
+
     if id != '-':
-        object = Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk') or request.form.get('ebuio_orgapk')), id=int(id)).first()
+        object = Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk') or request.form.get('ebuio_orgapk')),
+                                         id=int(id)).first()
 
     if request.method == 'POST':
 
         if not object:
             object = Picture(int(request.form.get('ebuio_orgapk')))
+
+        if sp:
+            object.image_url_prefix = sp.image_url_prefix
 
         object.name = request.form.get('name')
         object.radiotext = request.form.get('radiotext')
@@ -73,43 +88,61 @@ def radiovis_gallery_edit(request, id):
 
             return None
 
-        file = request.files['file']
-        if file:
-            filename = secure_filename(file.filename)
-            full_path = add_unique_postfix('media/uploads/radiovis/gallery/' + filename)
-            file.save(full_path)
-            if object.filename:
-                try:
-                    os.unlink(object.filename)
-                except:
-                    pass
-            object.filename = full_path
+        def unique_filename(fn):
 
-        # Check errors
-        if object.name == '':
-            errors.append("Please set a name")
+            path, name = os.path.split(fn)
+            name, ext = os.path.splitext(name)
 
-        if object.radiotext == '':
-            errors.append("Please set a text")
+            make_fn = lambda i: os.path.join(path, '%s%s' % (str(uuid.uuid4()), ext))
 
-        if object.radiolink == '':
-            errors.append("Please set a link")
+            for i in xrange(2, sys.maxint):
+                uni_fn = make_fn(i)
+                if not os.path.exists(uni_fn):
+                    return uni_fn
 
-        if object.filename == '' or object.filename is None:
-            errors.append("Please upload an image")
-        else:
+            return None
 
-            if imghdr.what(object.filename) not in ['jpeg', 'png']:
-                errors.append("Image is not an png or jpeg image")
-                os.unlink(object.filename)
-                object.filename = None
+        new_file = False
+        if request.files:
+            new_file = True
+            file = request.files['file']
+            if file:
+                filename = secure_filename(file.filename)
+                full_path = unique_filename('media/uploads/radiovis/gallery/' + filename)
+                path, name = os.path.split(full_path)
+                file.save(full_path)
+                if object.filename:
+                    try:
+                        os.unlink(object.filename)
+                    except:
+                        pass
+                object.filename = name
+
+            # Check errors
+            if object.name == '':
+                errors.append("Please set a name")
+
+            if object.radiotext == '':
+                errors.append("Please set a text")
+
+            if object.radiolink == '':
+                errors.append("Please set a link")
+
+            if object.filename == '' or object.filename is None:
+                errors.append("Please upload an image")
             else:
-                im = Image.open(object.filename)
-                if im.size != (320, 240):
-                    errors.append("Image must be 320x240")
-                    del im
-                    os.unlink(object.filename)
+
+                if imghdr.what(full_path) not in ['jpeg', 'png']:
+                    errors.append("Image is not an png or jpeg image")
+                    os.unlink(full_path)
                     object.filename = None
+                else:
+                    im = Image.open(full_path)
+                    if im.size != (320, 240):
+                        errors.append("Image must be 320x240")
+                        del im
+                        os.unlink(full_path)
+                        object.filename = None
 
         pieces = urlparse.urlparse(object.radiolink)
 
@@ -119,12 +152,35 @@ def radiovis_gallery_edit(request, id):
         # If no errors, save
         if not errors:
 
+            # Upload to s3
+            if new_file:
+                try:
+                    if config.STANDALONE:
+                        send_image_to_mock_api({name: open(full_path, 'rb')})
+                    else:
+                        awsutils.upload_public_image(sp, name, full_path)
+                except:
+                    pass
+            # TODO Clean up old images on S3
+
             if not object.id:
                 db.session.add(object)
 
             db.session.commit()
 
+            try:
+                # Remove local copy
+                os.unlink(full_path)
+            except:
+                pass
+
             return PlugItRedirect('radiovis/gallery/?saved=yes')
+
+        try:
+            # Remove local copy
+            os.unlink(full_path)
+        except:
+            pass
 
     if object:
         object = object.json
@@ -138,9 +194,25 @@ def radiovis_gallery_edit(request, id):
 def radiovis_gallery_delete(request, id):
     """Delete a picture."""
 
+    plugitapi = PlugItAPI(config.API_URL)
+    orga = plugitapi.get_orga(request.args.get('ebuio_orgapk') or request.form.get('ebuio_orgapk'))
+
+    sp = None
+    if orga.codops:
+        sp = ServiceProvider.query.filter_by(codops=orga.codops).order_by(ServiceProvider.codops).first()
+
     object = Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk')), id=int(id)).first()
 
-    os.unlink(object.filename)
+    # Remove File
+    try:
+        os.unlink(object.filename)
+    except:
+        pass
+    # Remove from S3
+    try:
+        awsutils.delete_public_image(sp, object.filename)
+    except:
+        pass
 
     db.session.delete(object)
     db.session.commit()
@@ -155,15 +227,46 @@ def radiovis_channels_home(request):
 
     list = []
 
-    for elem in Channel.query.join(Station).filter(Station.orga==int(request.args.get('ebuio_orgapk'))).order_by(Channel.name).all():
+    for elem in Channel.query.filter(Channel.type_id != 'id').join(Station).filter(
+            Station.orga == int(request.args.get('ebuio_orgapk'))).order_by(Station.name, Channel.type_id,
+                                                                            Channel.name).all():
         list.append(elem.json)
 
     pictures = []
 
-    for elem in Picture.query.filter_by(orga = int(request.args.get('ebuio_orgapk'))).order_by(Picture.name).all():
+    for elem in Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk'))).order_by(Picture.name).all():
         pictures.append(elem.json)
 
     return {'list': list, 'pictures': pictures}
+
+
+@action(route="/radiovis/channels/player/", template="radiovis/channels/player.html")
+@no_template()
+def radiovis_channels_player(request):
+    """Show a simple Player."""
+
+    topic = request.args.get('topic')
+    stationname = request.args.get('stationname')
+    name = request.args.get('name')
+
+    return {'stationname': stationname, 'name': name, 'topic': topic}
+
+
+@action(route="/radiovis/channels/playermodal/", template="radiovis/channels/playermodal.html")
+@no_template()
+def radiovis_channels_playermodal(request):
+    """Show a simple Player."""
+
+    topic = request.args.get('topic')
+    stationname = request.args.get('stationname')
+    name = request.args.get('name')
+
+    return {
+        'stationname': stationname,
+        'name': name, 'topic': topic,
+        'ws_endpoint': config.VIS_WEB_SOCKET_ENDPOINT_HOST,
+        'secure': not config.DEBUG
+    }
 
 
 @action(route="/radiovis/channels/set/<id>/<pictureid>")
@@ -172,7 +275,8 @@ def radiovis_channels_home(request):
 def radiovis_channels_set(request, id, pictureid):
     """Set a default value for a channel"""
 
-    object = Channel.query.join(Station).filter(Channel.id == int(id), Station.orga==int(request.args.get('ebuio_orgapk'))).first()
+    object = Channel.query.join(Station).filter(Channel.id == int(id),
+                                                Station.orga == int(request.args.get('ebuio_orgapk'))).first()
 
     picture = Picture.query.filter_by(orga=int(request.args.get('ebuio_orgapk')), id=int(pictureid)).first()
 
@@ -191,7 +295,8 @@ def radiovis_channels_set(request, id, pictureid):
 def radiovis_channels_logs(request, id):
     """Show logs for a channel"""
 
-    object = Channel.query.join(Station).filter(Channel.id == int(id), Station.orga==int(request.args.get('ebuio_orgapk'))).first()
+    object = Channel.query.join(Station).filter(Channel.id == int(id),
+                                                Station.orga == int(request.args.get('ebuio_orgapk'))).first()
 
     if not object:
         abort(404)
@@ -199,7 +304,8 @@ def radiovis_channels_logs(request, id):
 
     list = []
 
-    for logentry in LogEntry.query.filter(LogEntry.topic.ilike(object.topic + '%')).order_by(-LogEntry.reception_timestamp).all():
+    for logentry in LogEntry.query.filter(LogEntry.topic.ilike(object.topic + '%')).order_by(
+            -LogEntry.reception_timestamp).all():
         list.append(logentry.json)
 
     return {'list': list, 'channel': object.json}
@@ -300,6 +406,24 @@ def radiovis_api_get_all_channels(request, secret):
     return {'list': list}
 
 
+@action(route="/radiovis/api/<secret>/get_all_vis_channels")
+@only_orga_admin_user()  # To prevent call from IO
+@json_only()
+def radiovis_api_get_all_vis_channels(request, secret):
+    """Retrun the list of all channels"""
+
+    if secret != config.API_SECRET:
+        abort(404)
+        return
+
+    list = []
+
+    for channel in Channel.query.filter(Channel.default_picture_id != None).all():
+        list.append((channel.topic, channel.id))
+
+    return {'list': list}
+
+
 @action(route="/radiovis/api/<secret>/get_channel_default")
 @only_orga_admin_user()  # To prevent call from IO
 @json_only()
@@ -312,10 +436,8 @@ def radiovis_api_get_channel_default(request, secret):
 
     channel = Channel.query.filter_by(id=request.form.get('id')).first()
 
-    dp = channel.default_picture
-
-    if dp:
-        return {'info': dp.json}
+    if channel and channel.default_picture:
+        return {'info': channel.default_picture.json}
     else:
         return {'info': None}
 
@@ -337,7 +459,7 @@ def radiovis_api_cleanup_logs(request, secret):
     # db.session.commit()
     from sqlalchemy.sql import text
     db.engine.execute(text('DELETE FROM log_entry WHERE reception_timestamp <  :t'),
-                      t = (time.time() - int(request.form.get('max_age'))))
+                      t=(time.time() - int(request.form.get('max_age'))))
 
     return {}
